feat(4-3): FREE mode manual share specification — full-stack implementation + code review fixes
## Story 4.3: FREE Mode Manual Share Specification

### New Features
- **ExpenseFree domain class** (`split/domain/expenses/ExpenseFree.java`): sealed Expense
  subclass using proportional "parts" system. Parts are stored; amounts are calculated
  proportionally at read time (e.g., parts [2,3] on €100 → [€40, €60]).
- **POST /api/splits/{splitId}/expenses/free** endpoint: new type-specific REST endpoint
  accepting `{ amount, description, payerId, shares: [{participantId, parts}] }`.
- **AddFreeExpenseRequest** with Bean Validation (`@NotEmpty`, `@Valid List<ShareRequest>`).
- **ExpenseFreeDTO** + **ExpenseFreePersistenceDTO**: full DTO layers for API responses and
  JSON file persistence (parts stored, amounts calculated on load).
- **ExpenseEditModal.svelte**: unified add/edit modal replacing separate AddExpenseModal and
  EditExpenseModal. FREE mode adds participant share-parts inputs with running total and
  proportional calculation explanation.
- **ExpenseCard.svelte**: FREE mode displays "Manual split" in share breakdown.
- **splits.ts**: `addFreeExpense()` calling `/expenses/free`; `addExpense()` now routes to
  type-specific endpoints (`/expenses/by-night`, `/expenses/equal`) instead of deprecated
  generic `/expenses`.

### Domain Package Reorganisation
- `split/domain/expenses/` subpackage: Expense, ExpenseByNight, ExpenseEqual, ExpenseFree,
  SplitMode, ExpenseNotFoundError, InvalidExpenseIdError, InvalidSharesError, PayerNotFoundError.
- `split/domain/participant/` subpackage: Participant, DuplicateParticipantNameError,
  InvalidParticipantIdError, ParticipantHasExpensesError, ParticipantNotFoundError.
- `split/service/` package: AddExpenseRequest, AddFreeExpenseRequest, AddParticipantRequest,
  AddTypedExpenseRequest, CreateSplitRequest moved from split/domain/.
- `split/persistence/`: JsonFileRepository + TenantPathResolver moved from sharedkernel/;
  SplitRepository interface extracted.

### Code Review Fixes (story 4-3 review, 2026-02-11)
- **AC8 fix**: `SplitUseCases.addExpenseFree()` now validates that ALL split participants
  have a share specified — throws `InvalidSharesError("All participants must have a share
  specified. Missing share for participant '...'.")` when any participant is omitted.
- **AC8 test**: added `addExpenseFree_withMissingSplitParticipant_returns400` integration
  test to `ExpenseUseCaseTest.java` covering the correct AC8 scenario (valid participant
  present in split but absent from shares list).
- **Description max alignment**: `UpdateExpenseRequest` changed from 100→200 chars to match
  domain model (`Expense.MAX_DESCRIPTION_LENGTH=200`) and `AddFreeExpenseRequest`. Frontend
  `ExpenseEditModal` validation and error message updated to 200. Frontend tests updated.
- **Deprecated endpoint removed from frontend**: `addExpense()` in `splits.ts` now calls
  `/expenses/by-night` or `/expenses/equal` based on splitMode, removing dependency on the
  deprecated generic `/expenses` endpoint.

### Tests
- Backend: 10 unit tests in `ExpenseFreeTest.java` (parts validation, proportional calc,
  fromJson round-trip, edge cases).
- Backend: 9 FREE mode integration tests in `ExpenseUseCaseTest.java` (AC6-AC8, persistence
  round-trip, zero parts, negative parts, missing participant, valid participant omitted).
- Frontend: 14 tests in `AddExpenseModal.test.ts` for FREE mode (AC1-6, AC9).
- Frontend: 8 tests in `ExpenseCard.test.ts` for FREE display (AC10).

### Known Open Items (Review Follow-ups in story file)
- [MEDIUM] Rename AddExpenseModal.test.ts → ExpenseEditModal.test.ts
- [MEDIUM] FREE mode expense editing incomplete (isDirtyEdit/resetForm/updateExpense lack
  shares support — requires dedicated story)
- [LOW] SplitMaximumExpenseReachedError / SplitMaximumParticipantNumberReachedError need
  story assignment or removal
- [LOW] ExpenseCard total ✓ shown unconditionally (should verify shares sum)

+++ fairnsquare-app/src/test/java/org/asymetrik/web/fairnsquare/split/domain/expenses/ExpenseByNightTest.java
@@ -3,6 +3,10 @@ package org.asymetrik.web.fairnsquare.split.domain;
 import java.math.BigDecimal;
 import java.util.List;
 
+import org.asymetrik.web.fairnsquare.split.domain.expenses.Expense;
+import org.asymetrik.web.fairnsquare.split.domain.expenses.ExpenseByNight;
+import org.asymetrik.web.fairnsquare.split.domain.expenses.SplitMode;
+import org.asymetrik.web.fairnsquare.split.domain.participant.Participant;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -134,7 +138,7 @@ class ExpenseByNightTest {
         assertThat(expense.getPayerId()).isEqualTo(payerId);
     }
 
-    private Participant createParticipant(String name, int nights) {
+    private Participant createParticipant(String name, double nights) {
         return Participant.create(name, nights);
     }
 }
